pyproject.toml
[build-system]
requires = ["flit_core>=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "pytis"
description = "Pytis — Python framework for building information systems declaratively"
dynamic = ["version"]
authors = [
    { name = "Milan Zamazal", email = "pdm@zamazal.org" },
    { name = "Tomas Cerha", email = "cerha@truecode.cz" },
    { name = "Pavel Hanák", email = "pavel@hanak.name" },
]
license = { file = "COPYING" }
readme = "readme.md"
requires-python = ">=2.7"
classifiers = [
    "Programming Language :: Python :: 2.7",
    "Programming Language :: Python :: 3.5",
    "License :: OSI Approved :: GNU General Public License v2 or later (GPLv2+)",
]
dependencies = [
    "cachetools",
    "future",
    "gnupg",
    "psycopg2-binary",
    "sqlalchemy==1.4",
]

[project.optional-dependencies]
wx = [
    # Dependencies for Pytis wxPython desktop applications.
    "file-magic; python_version >= '3.0'",
    "lcg-framework[all]",
    "pillow>=2.7.0",
    "pycryptodome",
    "pymupdf==1.14.20; python_version == '2.7'",
    "pymupdf; python_version >= '3.0'",
    # If building pygobject fails, installing system libraries libcairo-dev and
    # libgirepository1.0-dev may help.  However pygobject >= 3.51 depends on
    # libgirepository2.0-dev, which is not yet available on stable distributions,
    # so fixing at 3.50 for now.
    "pygobject<=3.50; platform_system == 'Linux'",
    # RPyC 3.3.0 is needed in python 2.7, newest version works on Python 3
    "rpyc==3.3.0; python_version == '2.7'",
    "rpyc; python_version >= '3.0'",
    # To use a binary wxPython wheel, it may be necessary to add -f and limit 
    # wxPython version to the one available in given source, such as:
    # pip install -f https://extras.wxpython.org/wxPython4/extras/linux/gtk3/debian-10/ wxPython==4.2.1
    "wxpython",
]

[tool.flit.module]
name = "pytis"

[tool.flit.sdist]
include = [
    "COPYING",
    "pytis/translations/**/**.mo",
    "pytis/resources/**",
    "pytis/**/test.py",  # Excluded in .flitignore, so override for sdist.
]
